+++ src/SmartTyping.Domain/Enums/SettingKeys.cs
@@ -20,6 +20,7 @@ public static class SettingKeys
     public const string AutoCorrectSuggest = "AutoCorrectSuggest";
     public const string AutoCorrectAuto = "AutoCorrectAuto";
     public const string AutoExpandSnippets = "AutoExpandSnippets";
+    public const string ShowNotifications = "ShowNotifications";
     public const string HotkeyAiImprove = "HotkeyAiImprove";
     public const string AiProvider = "AiProvider";
     public const string AiApiKey = "AiApiKey";